+++ com.microsoft.java.test.runner/src/main/java/com/microsoft/java/test/runner/JUnitLauncher.java
@@ -16,21 +16,21 @@ public class JUnitLauncher {
     }
 
     private static int execute(String[] args) {
+        TestOutputStream stream = TestOutputStream.instance();
         try {
             if (args.length == 0) {
-                TestingMessageHelper.reporterAttached(System.out);
-                System.err.print("No test found to run");
+                TestingMessageHelper.reporterAttached(stream);
+                stream.println(new TestMessageItem("No test found to run", null));
             } else {
                 CustomizedJUnitCoreRunner jUnitCore = new CustomizedJUnitCoreRunner();
                 jUnitCore.run(args);
             }
             return 0;
         } catch (Throwable e) {
-            e.printStackTrace();
+            stream.println(new TestMessageItem("Failed to launch Junit", e));
             return 1;
         } finally {
-            System.err.flush();
-            System.out.flush();
+            stream.flush();
         }
     }
 }