+++ com.microsoft.java.test.runner/src/main/java/com/microsoft/java/test/runner/TestMessageType.java
@@ -1,0 +1,6 @@
+package com.microsoft.java.test.runner;
+
+public enum TestMessageType {
+    Info,
+    Error,
+}