+++ com.microsoft.java.test.runner/src/main/java/com/microsoft/java/test/runner/TestingMessageHelper.java
@@ -10,43 +10,24 @@
  */
 package com.microsoft.java.test.runner;
 
-import java.io.PrintStream;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.StringJoiner;
 
 import org.junit.runner.Description;
 import org.junit.runner.Result;
 import org.junit.runner.notification.Failure;
 
 public class TestingMessageHelper {
-    private static final String TEST_REPORTER_ATTACHED = "testReporterAttached";
-    private static final String ROOT_NAME = "rootName";
-    private static final String NAME = "name";
-    private static final String LOCATION = "location";
-    private static final String TEST_STARTED = "testStarted";
-    private static final String TEST_IGNORED = "testIgnored";
-    private static final String TEST_FINISHED = "testFinished";
-    private static final String DURATION = "duration";
-    private static final String SUITE_TREE_NODE = "suiteTreeNode";
-    private static final String TEST_SUITE_FINISHED = "testSuiteFinished";
-    private static final String TEST_SUITE_STARTED = "testSuiteStarted";
-    private static final String SUITE_TREE_STARTED = "suiteTreeStarted";
-    private static final String SUITE_TREE_ENDED = "suiteTreeEnded";
-    private static final String MESSAGE = "message";
-    private static final String DETAILS = "details";
-    private static final String TEST_FAILED = "testFailed";
-
     /**
      * Prints a message when the test reported was attached.
      *
-     * @param out output stream
+     * @param stream output stream
      */
-    public static void reporterAttached(PrintStream out) {
-        out.println(create(TEST_REPORTER_ATTACHED, (List<Pair>) null));
+    public static void reporterAttached(TestOutputStream stream) {
+        stream.println(create(TestMessageConstants.TEST_REPORTER_ATTACHED, (List<Pair>) null));
     }
 
     /**
@@ -54,8 +35,8 @@ public class TestingMessageHelper {
      *
      * @param out output stream
      */
-    public static void rootPresentation(PrintStream out) {
-        out.println(create(ROOT_NAME, new Pair(NAME, "Default Suite")));
+    public static void rootPresentation(TestOutputStream stream) {
+        stream.println(create(TestMessageConstants.ROOT_NAME, new Pair(TestMessageConstants.NAME, "Default Suite")));
     }
 
     /**
@@ -64,13 +45,13 @@ public class TestingMessageHelper {
      * @param description information about test
      * @param out output stream
      */
-    public static void testStarted(PrintStream out, Description description) {
+    public static void testStarted(TestOutputStream stream, Description description) {
         String location = description.getClassName() + "." + description.getMethodName();
-        out.println(
+        stream.println(
                 create(
-                        TEST_STARTED,
-                        new Pair(NAME, escape(description.getMethodName())),
-                        new Pair(LOCATION, "java:test://" + escape(location))));
+                        TestMessageConstants.TEST_STARTED,
+                        new Pair(TestMessageConstants.NAME, description.getMethodName()),
+                        new Pair(TestMessageConstants.LOCATION, "java:test://" + location)));
     }
 
     /**
@@ -79,8 +60,8 @@ public class TestingMessageHelper {
      * @param name method name
      * @param out output stream
      */
-    public static void testIgnored(PrintStream out, String name) {
-        out.println(create(TEST_IGNORED, new Pair(NAME, escape(name))));
+    public static void testIgnored(TestOutputStream stream, String name) {
+        stream.println(create(TestMessageConstants.TEST_IGNORED, new Pair(TestMessageConstants.NAME, name)));
     }
 
     /**
@@ -90,12 +71,12 @@ public class TestingMessageHelper {
      * @param out output stream
      * @param duration time of test running
      */
-    public static void testFinished(PrintStream out, Description description, long duration) {
-        out.println(
+    public static void testFinished(TestOutputStream stream, Description description, long duration) {
+        stream.println(
                 create(
-                        TEST_FINISHED,
-                        new Pair(NAME, escape(description.getMethodName())),
-                        new Pair(DURATION, String.valueOf(duration))));
+                        TestMessageConstants.TEST_FINISHED,
+                        new Pair(TestMessageConstants.NAME, description.getMethodName()),
+                        new Pair(TestMessageConstants.DURATION, String.valueOf(duration))));
     }
 
     /**
@@ -104,13 +85,13 @@ public class TestingMessageHelper {
      * @param description information about test node
      * @param out output stream
      */
-    public static void treeNode(PrintStream out, Description description) {
+    public static void treeNode(TestOutputStream stream, Description description) {
         String location = description.getClassName() + "." + description.getMethodName();
-        out.println(
+        stream.println(
                 create(
-                        SUITE_TREE_NODE,
-                        new Pair(NAME, escape(description.getMethodName())),
-                        new Pair(LOCATION, "java:test://" + escape(location))));
+                        TestMessageConstants.SUITE_TREE_NODE,
+                        new Pair(TestMessageConstants.NAME, description.getMethodName()),
+                        new Pair(TestMessageConstants.LOCATION, "java:test://" + location)));
     }
 
     /**
@@ -119,8 +100,8 @@ public class TestingMessageHelper {
      * @param currentSuite name of test suite
      * @param out output stream
      */
-    public static void testSuiteFinished(PrintStream out, String currentSuite) {
-        out.println(create(TEST_SUITE_FINISHED, new Pair(NAME, escape(currentSuite))));
+    public static void testSuiteFinished(TestOutputStream stream, String currentSuite) {
+        stream.println(create(TestMessageConstants.TEST_SUITE_FINISHED, new Pair(TestMessageConstants.NAME, currentSuite)));
     }
 
     /**
@@ -129,12 +110,12 @@ public class TestingMessageHelper {
      * @param description information about suite
      * @param out output stream
      */
-    public static void testSuiteStarted(PrintStream out, Description description) {
-        out.println(
+    public static void testSuiteStarted(TestOutputStream stream, Description description) {
+        stream.println(
                 create(
-                        TEST_SUITE_STARTED,
-                        new Pair(NAME, escape(description.getClassName())),
-                        new Pair(LOCATION, "java:test://" + escape(description.getClassName()))));
+                        TestMessageConstants.TEST_SUITE_STARTED,
+                        new Pair(TestMessageConstants.NAME, description.getClassName()),
+                        new Pair(TestMessageConstants.LOCATION, "java:test://" + description.getClassName())));
     }
 
     /**
@@ -143,12 +124,12 @@ public class TestingMessageHelper {
      * @param description information about suite
      * @param out output stream
      */
-    public static void suiteTreeNodeStarted(PrintStream out, Description description) {
-        out.println(
+    public static void suiteTreeNodeStarted(TestOutputStream stream, Description description) {
+        stream.println(
                 create(
-                        SUITE_TREE_STARTED,
-                        new Pair(NAME, escape(description.getClassName())),
-                        new Pair(LOCATION, "java:test://" + escape(description.getClassName()))));
+                        TestMessageConstants.SUITE_TREE_STARTED,
+                        new Pair(TestMessageConstants.NAME, description.getClassName()),
+                        new Pair(TestMessageConstants.LOCATION, "java:test://" + description.getClassName())));
     }
 
     /**
@@ -157,12 +138,12 @@ public class TestingMessageHelper {
      * @param description information about suite
      * @param out output stream
      */
-    public static void suiteTreeNodeEnded(PrintStream out, Description description) {
-        out.println(
+    public static void suiteTreeNodeEnded(TestOutputStream stream, Description description) {
+        stream.println(
                 create(
-                        SUITE_TREE_ENDED,
-                        new Pair(NAME, escape(description.getClassName())),
-                        new Pair(LOCATION, "java:test://" + escape(description.getClassName()))));
+                        TestMessageConstants.SUITE_TREE_ENDED,
+                        new Pair(TestMessageConstants.NAME, description.getClassName()),
+                        new Pair(TestMessageConstants.LOCATION, "java:test://" + description.getClassName())));
     }
 
     /**
@@ -172,9 +153,9 @@ public class TestingMessageHelper {
      * @param failure describes the test that failed and the exception that was thrown
      * @param duration time of test running
      */
-    public static void testFailed(PrintStream out, Failure failure, long duration) {
+    public static void testFailed(TestOutputStream stream, Failure failure, long duration) {
         List<Pair> attributes = new ArrayList<>();
-        attributes.add(new Pair(NAME, escape(failure.getDescription().getMethodName())));
+        attributes.add(new Pair(TestMessageConstants.NAME, failure.getDescription().getMethodName()));
         Throwable exception = failure.getException();
         if (exception != null) {
             String failMessage = failure.getMessage();
@@ -182,14 +163,14 @@ public class TestingMessageHelper {
             PrintWriter printWriter = new PrintWriter(writer);
             exception.printStackTrace(printWriter);
             String stackTrace = writer.getBuffer().toString();
-            attributes.add(new Pair(MESSAGE, escape(failMessage)));
-            attributes.add(new Pair(DETAILS, escape(stackTrace)));
+            attributes.add(new Pair(TestMessageConstants.MESSAGE, failMessage));
+            attributes.add(new Pair(TestMessageConstants.DETAILS, stackTrace));
         } else {
-            attributes.add(new Pair(MESSAGE, ""));
+            attributes.add(new Pair(TestMessageConstants.MESSAGE, ""));
         }
-        attributes.add(new Pair(DURATION, String.valueOf(duration)));
+        attributes.add(new Pair(TestMessageConstants.DURATION, String.valueOf(duration)));
 
-        out.println(create(TEST_FAILED, attributes));
+        stream.println(create(TestMessageConstants.TEST_FAILED, attributes));
     }
 
     /**
@@ -198,13 +179,15 @@ public class TestingMessageHelper {
      * @param out output stream
      * @param result the summary of the test run, including all the tests that failed
      */
-    public static void testRunFinished(PrintStream out, Result result) {
-        out.printf(
-                "Total tests run: %d, Failures: %d, Skips: %d",
+    public static void testRunFinished(TestOutputStream stream, Result result) {
+        String message = String.format("Total tests run: %d, Failures: %d, Skips: %d",
                 result.getRunCount(), result.getFailureCount(), result.getIgnoreCount());
+        stream.println(create(
+                TestMessageConstants.TEST_RESULT_SUMMARY,
+                new Pair(TestMessageConstants.MESSAGE, message)));
     }
 
-    private static String create(String name, Pair... attributes) {
+    private static TestMessageItem create(String name, Pair... attributes) {
         List<Pair> pairList = null;
         if (attributes != null) {
             pairList = Arrays.asList(attributes);
@@ -212,75 +195,8 @@ public class TestingMessageHelper {
         return create(name, pairList);
     }
 
-    private static String create(String name, List<Pair> attributes) {
-        StringBuilder builder = new StringBuilder("@@<{\"name\":");
-        builder.append('"').append(name).append('"');
-        if (attributes != null) {
-            builder.append(", \"attributes\":{");
-            StringJoiner joiner = new StringJoiner(", ");
-            for (Pair attribute : attributes) {
-                joiner.add("\"" + attribute.first + "\":\"" + attribute.second + "\"");
-            }
-            builder.append(joiner.toString());
-            builder.append("}");
-        }
-
-        builder.append("}>");
-        return builder.toString();
-    }
-    
-    private static String escape(String str) {
-    	if (str == null) {
-    		return str;
-    	}
-        int len = str.length();
-        StringBuilder sb = new StringBuilder(len);
-        String t;
-        for (int i = 0; i < len; i += 1) {
-            char c = str.charAt(i);
-            switch (c) {
-            case '\\':
-            case '\"':
-                sb.append('\\');
-                sb.append(c);
-                break;
-            case '\b':
-                sb.append("\\b");
-                break;
-            case '\t':
-                sb.append("\\t");
-                break;
-            case '\n':
-                sb.append("\\n");
-                break;
-            case '\f':
-                sb.append("\\f");
-                break;
-            case '\r':
-               sb.append("\\r");
-               break;
-            case '@':
-               sb.append("&#x40;");
-               break;
-            default:
-                if (c < ' ') {
-                    t = "000" + Integer.toHexString(c);
-                    sb.append("\\u" + t.substring(t.length() - 4));
-                } else {
-                    sb.append(c);
-                }
-            }
-        }
-        return sb.toString();
-    }
-
-    private static class Pair {
-        final String first;
-        final String second;
-
-        Pair(String first, String second) {
-            this.first = first;
-            this.second = second;
-        }
+    private static TestMessageItem create(String name, List<Pair> attributes) {
+        TestMessageItem item = new TestMessageItem(TestMessageType.Info, name, attributes);
+        return item;
     }
 }