+++ extension/src/Runner/JarFileRunner/jarFileRunnerResultAnalyzer.ts
@@ -9,5 +9,6 @@ export abstract class JarFileRunnerResultAnalyzer {
     constructor(protected _tests: ITestInfo[]) {
     }
     public abstract analyzeData(data: string): void;
+    public abstract analyzeError(error: string): void;
     public abstract feedBack(isCancelled: boolean): ITestResult[];
 }