+++ extension/src/projectManager.ts
@@ -1,41 +1,42 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+import * as path from 'path';
+import { workspace, CancellationToken, Uri } from 'vscode';
 import * as Commands from './Constants/commands';
 import * as Logger from './Utils/Logger/logger';
 
-import * as path from 'path';
-import { workspace, CancellationToken, Uri } from 'vscode';
-
 export class ProjectManager {
     // mapping from workspace folder uri to projects.
-    private projectInfos: Map<Uri, ProjectInfo[]> = new Map<Uri, ProjectInfo[]>();
+    private projectInfos: Map<string, ProjectInfo[]> = new Map<string, ProjectInfo[]>();
 
-    public async refresh(token?: CancellationToken): Promise<void[]> {
+    public async refresh(token?: CancellationToken): Promise<void> {
         if (!workspace.workspaceFolders) {
             return;
         }
-        return Promise.all(workspace.workspaceFolders.map((wkspace) => {
-            return this.getProjectInfo(wkspace.uri).then((infos: ProjectInfo[]) => {
-                infos.forEach((i) => { i.path = Uri.parse(i.path.toString()); });
-                this.storeProjects(wkspace.uri, infos);
-            },
-            (reason) => {
-                if (token.isCancellationRequested) {
+        await Promise.all(workspace.workspaceFolders.map(async (workspaceFolder) => {
+            try {
+                const infos: ProjectInfo[] = await this.getProjectInfo(workspaceFolder.uri);
+                for (const info of infos) {
+                    info.path = Uri.parse(info.path.toString());
+                }
+                this.storeProjects(workspaceFolder.uri, infos);
+            } catch (error) {
+                if (token && token.isCancellationRequested) {
                     return;
                 }
-                Logger.error(`Failed to refresh project mapping. Details: ${reason}.`);
-                return Promise.reject(reason);
-            });
+                Logger.error(`Failed to refresh project mapping. Details: ${error}.`);
+                throw error;
+            }
         }));
     }
 
     public storeProjects(wkspace: Uri, infos: ProjectInfo[]): void {
-        this.projectInfos.set(wkspace, infos);
+        this.projectInfos.set(wkspace.fsPath, infos);
     }
 
     public getProjects(wkspace: Uri): ProjectInfo[] {
-        return this.projectInfos.has(wkspace) ? this.projectInfos.get(wkspace) : [];
+        return this.projectInfos.has(wkspace.fsPath) ? this.projectInfos.get(wkspace.fsPath) : [];
     }
 
     public getAll(): ProjectInfo[] {
@@ -44,14 +45,15 @@ export class ProjectManager {
 
     public getProject(file: Uri): ProjectInfo {
         const fpath: string = this.formatPath(file.fsPath);
-        const matched = this.getAll().filter((p) => fpath.startsWith(this.formatPath(p.path.fsPath)));
+        const matched = this.getAll()
+                            .filter((p) => fpath.startsWith(this.formatPath(p.path.fsPath)))
+                            .sort((a, b) => (a.path.fsPath < b.path.fsPath ? 1 : -1));
         if (matched.length === 0) {
-            Logger.error(`Failed to get project.`);
+            Logger.error(`Failed to get the project.`);
             return undefined;
         }
         if (matched.length > 1) {
-            Logger.error(`Found multiple projects: ${matched.map((m) => m.name)}`);
-            return undefined;
+            Logger.warn(`Found multiple projects: ${matched.map((m) => m.name)}`);
         }
         return matched[0];
     }
@@ -77,8 +79,8 @@ export class ProjectManager {
         return formatted;
     }
 
-    private getProjectInfo(folder: Uri) {
-        return Commands.executeJavaLanguageServerCommand(Commands.JAVA_GET_PROJECT_INFO, folder.toString());
+    private getProjectInfo(folder: Uri): Thenable<ProjectInfo[]> {
+        return Commands.executeJavaLanguageServerCommand<ProjectInfo[]>(Commands.JAVA_GET_PROJECT_INFO, folder.toString());
     }
 }