+++ extension/src/testReportProvider.ts
@@ -1,13 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { TestResourceManager } from './testResourceManager';
-import { TestLevel, TestStatus, TestSuite } from './Models/protocols';
-
-import * as os from 'os';
 import * as path from 'path';
 import * as pug from 'pug';
-import { window, workspace, Event, EventEmitter, ExtensionContext, TextDocumentContentProvider, Uri, WorkspaceConfiguration } from 'vscode';
+import { workspace, Event, EventEmitter, ExtensionContext, TextDocumentContentProvider, Uri, WorkspaceConfiguration } from 'vscode';
+import { TestResourceManager } from './testResourceManager';
+import { TestLevel, TestStatus, TestSuite } from './Models/protocols';
 
 export class TestReportProvider implements TextDocumentContentProvider {