+++ extension/src/testResourceManager.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { commands, workspace, Event, EventEmitter, Uri } from 'vscode';
-import * as Commands from './Constants/commands';
+import { Event, EventEmitter, Uri } from 'vscode';
 import { Test, TestSuite } from './Models/protocols';
 import * as FetchTestsUtility from './Utils/fetchTestUtility';
 import * as Logger from './Utils/Logger/logger';
@@ -18,6 +17,9 @@ export class TestResourceManager {
         return this.testsIndexedByFileUri.has(path) ? this.testsIndexedByFileUri.get(path) : undefined;
     }
     public storeTests(file: Uri, tests: TestSuite[] | null | undefined): void {
+        if (tests === undefined || tests === null) {
+            return;
+        }
         const path = file.fsPath || '';
         const test: Test = {
             dirty: false,
@@ -25,6 +27,13 @@ export class TestResourceManager {
         };
         this.testsIndexedByFileUri.set(path, test);
         this._onDidChangeTestStorage.fire();
+    }
+    public removeTests(file: Uri): void {
+        const path = file.fsPath || '';
+        const deleted: boolean = this.testsIndexedByFileUri.delete(path);
+        if (deleted) {
+            this._onDidChangeTestStorage.fire();
+        }
     }
     public setDirty(file: Uri): void {
         const test = this.getTests(file);
@@ -38,7 +47,7 @@ export class TestResourceManager {
     }
     public getAll(): TestSuite[] {
         let allTests: TestSuite[] = [];
-        this.testsIndexedByFileUri.forEach((value, key, m) => {
+        this.testsIndexedByFileUri.forEach((value) => {
             allTests = allTests.concat(value.tests);
         });
         return allTests;
@@ -56,7 +65,7 @@ export class TestResourceManager {
                     collection.push(test);
                 }
             });
-            map.forEach((value, key, m) => {
+            map.forEach((value, key) => {
                 this.storeTests(Uri.parse(key), value);
             });
         },