+++ extension/src/testStatusBarProvider.ts
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 import { commands, window, ProgressLocation, StatusBarAlignment, StatusBarItem } from 'vscode';
-
 import * as Commands from './Constants/commands';
 import { TestLevel, TestStatus, TestSuite } from './Models/protocols';
 import { CommandUtility } from './Utils/commandUtility';
@@ -48,10 +47,9 @@ export class TestStatusBarProvider {
                 Logger.info('User canceled the long running operation');
                 commands.executeCommand(Commands.JAVA_TEST_CANCEL);
             });
-            p.report({ message: 'Running tests...', increment: 0 });
+            p.report({ message: 'Running tests...'});
             return action.then(() => {
                 this.updateStatus(tests);
-                p.report({ increment: 100 });
             },
             (reason) => {
                 this.statusBarItem.text = 'Failed to run tests';